adds analysis setups

+++ analysis/html-analysis.sh
@@ -42,15 +42,22 @@ function html_summary {
 	html_summary_image $1 DegreeDistributionR.degreeMax
 	html_summary_image $1 DegreeDistributionR.degreeMin
 	
-	# html_summary_image $1 UndirectedClusteringCoefficientU.averageCC
-	# html_summary_image $1 UndirectedClusteringCoefficientU.globalCC
+	html_summary_image $1 UndirectedClusteringCoefficientU.averageCC
+	html_summary_image $1 UndirectedClusteringCoefficientU.globalCC
 	
-	# html_summary_image $1 UnweightedAllPairsShortestPathsR.characteristicPathLength
-	# html_summary_image $1 UnweightedAllPairsShortestPathsR.connectivity
-	# html_summary_image $1 UnweightedAllPairsShortestPathsR.diameter
-	# html_summary_image $1 UnweightedAllPairsShortestPathsR.possiblePaths
+	html_summary_image $1 UnweightedAllPairsShortestPathsR.characteristicPathLength
+	html_summary_image $1 UnweightedAllPairsShortestPathsR.connectivity
+	html_summary_image $1 UnweightedAllPairsShortestPathsR.diameter
+	html_summary_image $1 UnweightedAllPairsShortestPathsR.possiblePaths
 
-	# html_summary_image $1 AssortativityR-out-unweighted.AssortativityCoefficient
+	html_summary_image $1 AssortativityR-out-unweighted.AssortativityCoefficient
+
+	html_summary_image $1 UndirectedMotifsU.UM1
+	html_summary_image $1 UndirectedMotifsU.UM2
+	html_summary_image $1 UndirectedMotifsU.UM3
+	html_summary_image $1 UndirectedMotifsU.UM4
+	html_summary_image $1 UndirectedMotifsU.UM5
+	html_summary_image $1 UndirectedMotifsU.UM6
 
 	html_footer
 
@@ -59,16 +66,33 @@ function html_summary {
 	done
 }
 
-function html_summary_image {
-	echo "<h2>$2</h2>"
-	echo "<div id='' style='overflow-y:scroll; height:auto; border:1px dotted;'>"
-	echo "<div id='' style='width:10000px;'>"
+function plot_exists {
 	for graph in $(ls $outputDir/$1 | grep -v 'index.'); do
-		echo "<a href='$graph/$2.png'><img width='350' src='$graph/$2.png' alt='$graph' title='$graph'/></a>"
+		if [[ -e $outputDir/$1/$graph/$2.png ]]; then
+			echo "1"
+			return
+		fi
 	done
-	echo "</div>"
-	echo "</div>"
-	echo "<?php spacer(); ?>"
+	echo "0"
+}
+
+function html_summary_image {
+	if [[ $(plot_exists $1 $2) -eq "1" ]]; then
+		echo "<h2>$2</h2>"
+		echo "<div id='' style='overflow-y:scroll; height:auto; border:1px dotted;'>"
+		width=$(($(ls $outputDir/$1 | grep -v 'index.' | wc -l) * 360))
+		echo "<div id='' style='width:${width}px;'>"
+		for graph in $(ls $outputDir/$1 | grep -v 'index.'); do
+			if [[ -e "$outputDir/$1/$graph/$2.png" ]]; then
+				echo "<a href='$graph/$2.png'><img width='350' src='$graph/$2.png' alt='$graph' title='$graph'/></a>"
+			else
+				echo "<img width='350' src='/img/etc/none.png' alt='no result for $graph' title='no result for $graph'/>"
+			fi
+		done
+		echo "</div>"
+		echo "</div>"
+		echo "<?php spacer(); ?>"
+	fi
 }
 
 function html_single {
@@ -78,4 +102,10 @@ function html_single {
 }
 
 outputDir="../results-mcnc"
+html_list > $outputDir/index.php
+
+outputDir="../results-properties"
+html_list > $outputDir/index.php
+
+outputDir="../results-walking-type"
 html_list > $outputDir/index.php
